+++ rest/src/main/java/de/starwit/rest/controller/RecordingController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import de.starwit.service.impl.StreamSavingService;
+import de.starwit.service.streamprocessing.StreamSavingService;
 import io.swagger.v3.oas.annotations.Operation;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
