+++ service/src/main/java/de/starwit/service/streamprocessing/StreamAvailabilityService.java
@@ -1,4 +1,4 @@
-package de.starwit.service.impl;
+package de.starwit.service.streamprocessing;
 
 import java.time.Duration;
 import java.util.Arrays;
